Move ruler code from VerticalStrip to LadderPage
VerticalStrip had info about optional markings that where not
essential. LadderPage(s) had in turn that same info copied over
and over.

VerticalStrip is now only a mapping from a numerical range to
an image. The optional markings are now handled by subclasses,
i.e LadderPage which stores the info in their descriptor thus
avoiding duplication. That could be made even more generic with
a Ruler class that would work hand-in-hand with VerticalStrips
and have LadderPageDescriptor have a Ruler member.