Refactor BufferedGauge/Animated Gauge into BaseGauge
Big refactor that must be carried on at once.
-Gauges are now stateful, the rendering is divided in two steps
(two virtual functions to implement), update_state and render.
*update_state allow gauges to do and save the computational work
that ultimaltely leads to which parts of which surfaces/textures
must be blit where(in the gauge's space).
*render tells the gauge to actually draw its content somewhere,
most of the time using the data set during update_state. At this
stage the gauge is supposed to only do drawing calls on pre-determined
coordinates/values.
=>This replaces the renderqueue/buffer.
-Animations are decoupled from gauges. The new BaseAnimation class
can now animate any (float) value(s) given their adresses. A single
animation can this way animate the values of several gauges at once.
-BaseGauge manages the rendering, advances animations - dirtying the
gauge has needed - and provides gfx calls instead of going through
AnimatedGauge and BufferedGauge.
It will also trigger a call to the gauge's update_state function if
the gauge is dirty.
-Any gauge can have children and be animated.
-Gauge buffering has not be ported over but is trivial to implement
if the need arises.
=>Overall 20% speed increase despite the recursive base_gauge_render
calls.