Refactor step1 - OdoGauge working again - Squash me later
BaseGauge is now responsible for drawing. Animations are not
handled by deriving BaseGauge anymore. BaseGauge has a frame
(that can be animated later on) and can have children.
Gauges are now stateful and must retain their state. Two virtual
functions: update_state (compute coordinates and retain sources)
and render(use state to draw data) to a location. BaseGauge triggers
the update_state when the gauge is dirty or if an animation is running.

BufferedGauge not used anymore. Less classes less virtual methods
indirection get rid of the queue.