Refactor gauge dispose/free pattern
BaseGauge now takes care of freeing child gauges, eliminating the
need to have per-gauge dispose() and free() functions dedicated
to that task.

BaseGaugeOps now provide a dispose field that gauges can use to
register a hook if they need to free resources they have allocated
themselves like SDL_Surfaces, char*, etc. Note that if you derive
a gauge that has provided such a hook and you need to register a
hook as well, you'll need to chain-up to the parent using
self->super.ops->dispose (or call directly the dispose function of
the derived class if it's not static.).

_dispose and _free functions now always return respectively self
and NULL as convenience that allows to write shorter failure
management code such as "return base_gauge_free(self);" that will
take care of freeing resources and return NULL to indicate the
failure.