Add property sets and listeners to DataSource
DataSource can now be (is) used to synchronize properties,
through property sets: LocationData, AttitudeData, etc.
We register listeners (instance+method) that will be
triggered when another piece of the code sets the property.

This puts the previous "glue" out of the main loop and lays
foundation for the future Route/Waypoints/etc handling between
the dialogs and other parts of the app (i.e map, header, etc.)