Replace base_gauge_dispose by base_gauge_free in _new functions
The following gauges *_init functions do not call base_gauge_init()
as their first operation: CompasGauge, AltGroup, ElevatorGauge
FishboneGauge, OdoGauge, VerticalStair. This can lead to a leak
if that gauge had to register a dispose function and fails before
registering it with a base_gauge_init call.

This is often because height and width are not known before doing
other operations that can fail. To address that, we have to modfify
BaseGauge to allow initing it without height/width and set those
values afterwise.